feat: 修改ws默认请求及调用入口Demo
