+++ asyncapi/client/index.html
@@ -11,6 +11,6 @@
     <div id="root">
       <h1>WebSocket Demo</h1>
     </div>
-    <script src="./index.js"></script>
+    <script type="module" src="./js/index.js"></script>
   </body>
 </html>
