+++ asyncapi/client/package.json
@@ -3,11 +3,13 @@
   "version": "1.0.0",
   "description": "",
   "scripts": {
-    "dev": "npx http-server -c-1 --cors='*'",
+    "dev": "npx vite",
     "test": "echo \"Error: no test specified\" && exit 1"
   },
   "keywords": [],
   "author": "Miles",
   "license": "ISC",
-  "devDependencies": {}
+  "devDependencies": {
+    "vite": "^4.3.9"
+  }
 }
